emit & end_head0 call new begin_body_if_necessary
A slight change in approach, in order to get ordering of output
correct.

The logic for outputting the title, closing head, and beginning
body have now moved into a new method begin_body_if_necessary().

Both emit() and end_head0() now call this method.

We now use a flag 'in_html_head' rather than 'body_tags_emitted'
for consistency with other in_* flags.

Note that we do everything but actually output the title tag at
this point.

Also that because of how emit already works, there is now always
a blank line before </head> in the output. This is the only
externally-visible change.